Removes the service.crunchService.* permission from basic users, as it allowed all users access to elevated Crunch permissions—such as granting themselves arbitrary capabilities without proper validation, and altering the capabilities of other users.
As a result of this removal, we now require a method in crunchService to pass a different user in context (one with elevated permissions) to perform these operations when expected—for example, when the system or ops user grants a grantMode: MANUAL capability to a basic user. The updateUserJunction method has been updated accordingly: it no longer uses Auth.sudo to switch into the UCJ source user's context, and instead maintains the context of the user requesting the action from the Crunch service. The updateJunctionFor method should now be used when Auth.sudo behavior is required for operations on ucjDAO.